Former-commit-id: 9a81e1f68b0c085ca87e8c16cf61f1ef7d28e4d6

+++ config/tool/properties/build_config_files.sh
@@ -9,22 +9,26 @@ function print_banner() {
     echo ""
 }
 
-if [[ `dirname $0` != './config/tool/properties' ]] ; then
-    print_banner
-    echo 'NO me ejecutes desde aquí que me electrocuto'
+function print_use() {
     echo ""
     echo "Uso:"
     echo "      ./config/tool/properties/build_config_files.sh <cliente> <entorno>"
     echo ""
+    echo "Parámetros:"
+    echo "      <entorno>: desa, pre, pro01, pro02, etc."
+    echo ""
+}
+
+if [[ `dirname $0` != './config/tool/properties' ]] ; then
+    print_banner
+    echo 'NO me ejecutes desde aquí que me electrocuto'
+    print_use
     exit 1
 fi
 
 if [ "$#" -lt 2 ]; then
     print_banner
-    echo ""
-    echo "Uso:"
-    echo "      ./config/tool/properties/build_config_files.sh <cliente> <entorno>"
-    echo ""
+    print_use
     exit 1
 fi 
 
@@ -32,16 +36,40 @@ print_banner
 cliente=$1
 entorno=$2
 
+
+if [[ ! -f config/batch/${cliente}/devon.properties ]] ; then
+    echo ""
+    echo "No existe el fichero:" config/batch/${cliente}/devon.properties
+    echo ""
+    exit 1
+fi
+
+if [[ ! -f config/web/${cliente}/devon.properties ]] ; then
+    echo ""
+    echo "No existe el fichero:" config/web/${cliente}/devon.properties
+    echo ""
+    exit 1
+fi
+
+if [[ ! -f ./config/tool/properties/${cliente}/${entorno}.properties ]] ; then
+    echo ""
+    echo "No existe el fichero:" ./config/tool/properties/${cliente}/${entorno}.properties
+    echo ""
+    exit 1
+fi
+
 mkdir -p config/tmp/batch/$cliente/
 mkdir -p config/tmp/web/$cliente/
+
 cp config/batch/${cliente}/devon.properties config/tmp/batch/$cliente/devon.properties
 cp config/web/${cliente}/devon.properties config/tmp/web/$cliente/devon.properties
+
+source ./config/tool/properties/${cliente}/${entorno}.properties
 
 while read line; do
     if [[ $line =~ ^.+=.+$ ]] ; then
         KEY=`echo $line | cut -d= -f1`
-        VALUE=`echo $line | cut -d= -f2`
-        VALUE=${VALUE//\//\\/}
+        VALUE=${!KEY//\//\\/}
         sed -e "s/${KEY}/${VALUE}/g" -i config/tmp/batch/${cliente}/devon.properties
         sed -e "s/${KEY}/${VALUE}/g" -i config/tmp/web/${cliente}/devon.properties
     fi
